Create SQL Project - BellaBeats
This project was completed as part of the Google Data Analytics Capstones. I used the 'Activity', 'Calories', 'Intensities', and 'Steps' tables from this public dataset. https://www.kaggle.com/arashnic/fitbit

Though this dataset was limited by only 33 users that wore the devices for a month, there was good info here. As far as overall trends, I noticed a wide range of user activity in the data. But the data did show the obvious positive correlation between total steps and calories burned. We also learned that calories burned and total active minutes does not correlate. 

Recommendations
- Promote BellaBeats as the fitness tracker for all types of exercise
- Include a variety or workout recommendations based on intensity and time preferences of the user